Fix issue with .d.ts files
The import path for the new `ApolloMockedProviderOptions` was not properly
prefixed with `./` which caused issues when using the package in a
typescript project.

Also, added the example in the README because last time I forgot to
`git add` it.

And last but not least, bumped version number. Hope that is ok.